fix: remove memory allocations in interrupts
Also removes *possible* divisions by zero.

This commit also removes the nmea parser and replaces it with a fairly junky ingrown solution. The old nmea parser allocated a shit ton of memory inside the GPS UART IRQ, leading to entire core hangs. It was not fun to debug, but thank god we have gdb.